i am stupid yes i am
L in L(P) => cover or incomp
not(cover or incomp) => not(L in L(P))
Given we know not(L in L(P)), not(cover or incomp) can be T or F
so it gets us nothing.
If we fix not(cover or incomp), we indeed get not(L in L(P)), but it's
too strong. It constrains more than we ask for, because it could be the
case that not(L in L(P)) and (cover or incomp).

=> we need find a p such that not(L in L(P)) => p
so that when we are given not(L in L(P)), we encode p

=> sufficient condition is too much; we need the necessary condition